requirements.txt
# ===== Development =====

pypdf==6.4.1
langchain-text-splitters==1.0.0

# ===== Production=====
fastapi==0.124.2
uvicorn==0.38.0
python-multipart==0.0.20
langchain==1.1.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.3
langchain-google-genai==4.0.0
langchain-pinecone==0.2.13
pinecone==7.3.0
sentence-transformers==5.1.2
python-dotenv==1.2.1
google-generativeai==0.8.5
-e .

